Reduce warnings on windows.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@28220 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Filters/avtSubsetBlockMergeFilter.C
@@ -224,7 +224,7 @@ avtSubsetBlockMergeFilter::Execute()
     delete [] appender;
     delete [] outDataSets;
 
-    for(size_t i=0; i<outSize; i++)
+    for(int i=0; i<outSize; i++)
     {
         cleaner[i]->Delete();
     }
